Created classes for different responsabilities. EKIllumination properties marked as public as there's no sense to have a public illuminance with only internal properties.
