+++ Package.resolved
@@ -6,8 +6,8 @@
         "repositoryURL": "https://github.com/ESKARIA/ESDateHelper.git",
         "state": {
           "branch": null,
-          "revision": "ca2c69f4c6f72d342fa7bea3a4a1efe58bd7c825",
-          "version": "1.0.0"
+          "revision": "c6c6af34536f3afd3996b6a922f66e2bea0479ab",
+          "version": "1.1.0"
         }
       }
     ]
